refactor: remove some leftovers
