--- app/lib/verifyCredentials.ts
@@ -1,85 +1,0 @@
-import {Credential, CredentialError } from '../types/credential';
-import {RegistryClient} from '@digitalcredentials/issuer-registry-client';
-import {extractCredentialsFrom, } from './verifiableObject';
-import vc from '@digitalcredentials/vc';
-import {checkStatus} from '@digitalcredentials/vc-status-list';
-import type {VerifyResponse} from './validate';
-import {securityLoader} from '@digitalcredentials/security-document-loader';
-import {Ed25519Signature2020} from '@digitalcredentials/ed25519-signature-2020';
-import {purposes} from '@digitalcredentials/jsonld-signatures';
-import {issuerInRegistries} from './issuerInRegistries';
-import {PresentationError, VerifiablePresentation} from '../types/presentation';
-
-const documentLoader = securityLoader({ fetchRemoteContexts: true }).build();
-const suite = new Ed25519Signature2020();
-const presentationPurpose = new purposes.AssertionProofPurpose();
-
-export async function verifyPresentation(
-  presentation: VerifiablePresentation,
-  unsignedPresentation = true,
-): Promise<VerifyResponse> {
-  try {
-    const hasRevocation = extractCredentialsFrom(presentation)?.find(
-      vc => vc.credentialStatus);
-    const result = await vc.verify({
-      presentation,
-      presentationPurpose,
-      suite,
-      documentLoader,
-      unsignedPresentation,
-      // Only check revocation status if any VC has a 'credentialStatus' property
-      checkStatus: hasRevocation ? checkStatus : undefined
-    });
-
-    if (!result.verified) {
-      console.warn('VP not verified:', JSON.stringify(result, null, 2));
-    }
-    return result;
-  } catch (err) {
-    console.warn(err);
-
-    throw new Error(PresentationError.CouldNotBeVerified);
-  }
-}
-
-
-
-export async function verifyCredential(credential: Credential, registries: RegistryClient): Promise<VerifyResponse> {
-  const { issuer } = credential;
-  const isInRegistry = issuerInRegistries({ issuer, registries });
-  if (!isInRegistry) {
-    throw new Error(CredentialError.DidNotInRegistry);
-  }
-
-  try {
-    const hasStatusProperty = extractCredentialsFrom(credential)?.find(
-      vc => vc.credentialStatus);
-    const result = await vc.verifyCredential({
-      credential,
-      suite,
-      documentLoader,
-      // Only check revocation status if VC has a 'credentialStatus' property
-      checkStatus: hasStatusProperty ? checkStatus : undefined,
-    });
-
-    // This logic catches the case where the verify response does not contain a `log` value
-    if (result.results?.[0].log === undefined) {
-      throw result.error || new Error('Verify response does not a `log` value');
-    }
-
-    if (!result.verified) {
-      console.warn('VC not verified:', JSON.stringify(result, null, 2));
-    }
-
-    return result;
-  } catch (err) {
-    console.warn(err);
-    console.log(JSON.stringify(err, removeStackReplacer, 2));
-
-    throw new Error(CredentialError.CouldNotBeVerified);
-  }
-}
-
-function removeStackReplacer(key: string, value: unknown) {
-  return key === 'stack' ? '...' : value;
-}
